# docs/swagger.yaml
basePath: /api
definitions:
...
  application.CreateRequest:
    properties:
      ad_id:
        type: string
      company_id:
        type: string
      influencer_id:
        type: string
    required:
    - ad_id
    - company_id
    - influencer_id
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
  description: API Server for Marketplace for influences and companies
  title: Marketplace
  version: "1.0"
paths:
...
  /application/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve an application based on its unique ID
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get an application by its ID
      tags:
      - application
  /application/{id}/delete:
    delete:
      consumes:
      - application/json
      description: Delete the application based on its ID
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete an application by ID
      tags:
      - application
  /application/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of an application by its ID
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: status
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update the status of an application
      tags:
      - application
  /application/ad/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve applications based on the Ad ID
      parameters:
      - description: Ad ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get applications by Ad ID
      tags:
      - application
  /application/create:
    post:
      consumes:
      - application/json
      description: Create a new application with the provided data
      parameters:
      - description: Application Create Request
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/application.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create an application
      tags:
      - application
  /application/influencer/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve applications based on the Influencer ID
      parameters:
      - description: Influencer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get applications by Influencer ID
      tags:
      - application
...
securityDefinitions:
  BearerAuth:
    description: JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"